ui/pyproject.toml
[project]
name = "ai-tools"
version = "0.1.0"
description = ""
authors = [
    {name = "Nguyen Huu Tho",email = "huu.thonguyen76@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "streamlit (>=1.51.0,<2.0.0)",
    "requests (>=2.32.5,<3.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
